App config moved to models
