+++ src/app/features/app/app.spec.ts
@@ -1,7 +1,7 @@
 import { it, describe, expect, beforeEach } from '@angular/core/testing';
 import { AppComponent } from './app.component';
 import { ConfigurationService } from '../../configuration-service';
-import { AppConfig } from '../../app-config';
+import { AppConfig } from '../../core/models/app-config';
 
 describe('AppComponent', () => {
 	let appComponent: AppComponent;
